+++ LanturnGame/main.cpp
@@ -30,8 +30,8 @@ int WINAPI WinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance, _
 
 
 	Sounds::LoadSounds();
-	Sounds::SetSoundVolume(Sounds::BGM, UserData::LoadData(UserData::SOUNDS, 0));
-	Sounds::SetSoundVolume(Sounds::SE, UserData::LoadData(UserData::SOUNDS, 1));
+	Sounds::SetSoundVolume(Sounds::BGM, (int)UserData::LoadData(UserData::SOUNDS, 0));
+	Sounds::SetSoundVolume(Sounds::SE, (int)UserData::LoadData(UserData::SOUNDS, 1));
 
 	SceneManager scene = (dynamic_cast<AbstractScene*>(new GameMain()));
